Update dependencies for improved compatibility and security

+++ crates/oxify-connect-storage/src/providers/azure_blob.rs
@@ -6,7 +6,7 @@ use std::sync::Arc;
 use std::time::Duration;
 
 use bytes::Bytes;
-use object_store::{azure::MicrosoftAzureBuilder, path::Path as StoragePath, ObjectStore};
+use object_store::{azure::MicrosoftAzureBuilder, path::Path as StoragePath, ObjectStore, ObjectStoreExt};
 use tracing::{debug, instrument};
 
 use super::ObjectStoreProvider;
@@ -194,7 +194,7 @@ impl ObjectStoreProvider for AzureBlobStoreProvider {
             let meta = item.map_err(|e| StorageError::Provider(e.to_string()))?;
             results.push(ObjectListing {
                 key: meta.location.to_string(),
-                size: meta.size as u64,
+                size: meta.size,
                 last_modified: Some(meta.last_modified),
                 etag: meta.e_tag,
             });
